+++ example/lib/pages/custom.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_map/flutter_map.dart';
-import 'package:latlong/latlong.dart';
 import 'package:flutter_map_location/flutter_map_location.dart';
 
 import '../widgets/drawer.dart';
@@ -40,14 +39,14 @@ class _CustomPageState extends State<CustomPage> {
               MarkerLayerOptions(markers: userLocationMarkers),
               LocationOptions(
                 markers: userLocationMarkers,
-                onLocationUpdate: (LatLng loc) {
-                  print('Location updated: $loc');
+                onLocationUpdate: (LatLngData ld) {
+                  print('Location updated: ${ld?.location}');
                 },
-                onLocationRequested: (LatLng loc) {
-                  if (loc == null) {
+                onLocationRequested: (LatLngData ld) {
+                  if (ld == null || ld.location == null) {
                     return;
                   }
-                  mapController?.move(loc, 16.0);
+                  mapController?.move(ld.location, 16.0);
                 },
                 buttonBuilder: (BuildContext context,
                     ValueNotifier<LocationServiceStatus> status,
@@ -66,10 +65,10 @@ class _CustomPageState extends State<CustomPage> {
                     ),
                   );
                 },
-                markerBuilder: (BuildContext context, LatLng point,
+                markerBuilder: (BuildContext context, LatLngData ld,
                     ValueNotifier<double> heading) {
                   return Marker(
-                    point: point,
+                    point: ld.location,
                     builder: (_) => Container(
                       child: Column(
                         children: <Widget>[