+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Dev
+
+* Add possiblity to change marker icon depending on the location accuracy.
+* BREAKING CHANGE: All options return the type `LatLngData` instead of `LatLng`.
+
 ## 0.6.0
 
 * Update flutter_compass to allow usage of geomagnetic rotation sensor as fallback on Android.